requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.0
python-dotenv==1.0.0
gunicorn==23.0.0
certifi~=2024.12.14
google-apps-meet==0.1.15
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.165.0
flask-bcrypt==1.0.1
flask-login==0.6.3
Flask-WTF==1.2.2
wtforms~=3.2.1
apscheduler==3.11.0
boto3~=1.36.7
openai~=1.61.0
psycopg2-binary==2.9.10
werkzeug~=3.1.3
alembic~=1.14.0
SQLAlchemy~=2.0.37
flask-jwt-extended==4.7.1
requests~=2.32.3
Flask-Session==0.8.0
redis~=5.2.1
flask-session==0.8.0
celery==5.4.0
